+++ chat-widget/src/components/footerstateful/FooterStateful.tsx
@@ -1,5 +1,5 @@
 import { LogLevel, TelemetryEvent } from "../../common/telemetry/TelemetryConstants";
-import React, { Dispatch } from "react";
+import React, { Dispatch, useEffect } from "react";
 
 import AudioNotificationStateful from "./audionotificationstateful/AudioNotificationStateful";
 import { Constants } from "../../common/Constants";
@@ -62,6 +62,12 @@ export const FooterStateful = (props: any) => {
             isAudioMuted: state.appStates.isAudioMuted
         },
     };
+
+    useEffect(() => {
+        if (state.appStates.isAudioMuted === null) {
+            dispatch({ type: LiveChatWidgetActionType.SET_AUDIO_NOTIFICATION, payload: footerProps?.controlProps?.audioNotificationButtonProps?.isAudioMuted ?? false });
+        }
+    }, []);
 
     return (
         <>